backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.1
python-decouple==3.8
PyJWT==2.8.1
bcrypt==4.1.1
mysqlclient==2.2.0
python-dateutil==2.8.2
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
celery==5.3.1
redis==4.5.5
gunicorn==20.1.0
